+++ js/components/common/react_modal.js.jsx
@@ -2,7 +2,7 @@ import React, { createRef } from "react"
 import ReactModal from "react-modal"
 import { FeatureGroup, MapContainer, Marker, Popup, TileLayer } from "react-leaflet"
 import L from "leaflet"
-import { getCourse, getPost } from "../common/utils"
+import { getCourse, getProgram } from "../common/utils"
 import { AgGridReact } from "ag-grid-react"
 
 const DAY_TO_INT = {
@@ -159,7 +159,7 @@ class CourseModal extends React.Component {
         occurrences.rooms.push(firstRoom + secondRoom)
         occurrences.times.push(
           DAY_TO_INT[occurrence.weekDay] +
-            "  " +
+            " " +
             occurrence.startHour +
             " - " +
             occurrence.endHour
@@ -266,28 +266,43 @@ class Description extends React.Component {
       timetableDisplay = Object.keys(this.props.sessions).map(session =>
         this.props.sessions[session].length !== 0 ? (
           <div key={session}>
-            <strong>{this.props.course.name + "-" + session}</strong>
-            <div className="ag-theme-alpine" style={{ height: 500, width: 940 }}>
+            <strong className="semester-heading">
+              {session === "F" ? "Fall" : session === "S" ? "Winter" : "Full Year"}
+            </strong>
+            <div className="ag-theme-alpine" style={{ width: 940 }}>
               <AgGridReact
                 rowData={this.props.sessions[session]}
                 columnDefs={[
                   { field: "activity", width: 130 },
-                  { field: "instructor", width: 170 },
-                  { field: "availability" },
-                  { field: "waitList", width: 120 },
+                  { field: "instructor", width: 190 },
+                  { field: "availability", width: 180 },
+                  { field: "waitList", width: 130 },
                   {
                     field: "time",
-                    cellStyle: { whiteSpace: "pre" },
+                    cellStyle: {
+                      whiteSpace: "pre",
+                      lineHeight: "1.8",
+                      paddingTop: "7px",
+                      paddingBottom: "6px",
+                    },
                     valueFormatter: col => col.data.time.join("\n"),
+                    width: 180,
+                    autoHeight: true,
                   },
                   {
                     field: "room",
-                    cellStyle: { whiteSpace: "pre" },
+                    cellStyle: {
+                      whiteSpace: "pre",
+                      lineHeight: "1.8",
+                      paddingTop: "7px",
+                      paddingBottom: "6px",
+                    },
                     valueFormatter: col => col.data.room.join("\n"),
-                    width: 140,
+                    width: 128,
+                    autoHeight: true,
                   },
                 ]}
-                rowHeight={100}
+                domLayout="autoHeight"
               ></AgGridReact>
             </div>
           </div>
@@ -365,7 +380,7 @@ class FocusModal extends React.Component {
 
   componentDidUpdate(prevProps) {
     if (this.props.showFocusModal && !prevProps.showFocusModal) {
-      getPost(this.props.focusId, this.state.focusModifiedTime).then(focusData => {
+      getProgram(this.props.focusId, this.state.focusModifiedTime).then(focusData => {
         if (!focusData.modified) return
 
         this.setState({